+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Convenience constructor for creating search spaces from dataframes
+
+### Changed
+- Cached recommendations are now a private attribute
+- Parameters, targets and objectives are now immutable
+
 ## [0.3.1] - 2023-07-17
 ### Added
 - Code coverage check with pytest-cov