+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Constraints serialization
 
+### Changed
+- A maxiumum of one `DependenciesConstraint` is allowed
+
 ## [0.3.1] - 2023-07-17
 ### Added
 - Code coverage check with pytest-cov